[crypto] Implement key integrity checksums.
Uses a CRC32 checksum to provide a lightweight integrity check on keys. The
basic setup for this already existed, this commit just:
- fills in the stub
- corrects a few places that didn't write the checksum
- changes a couple fields in the key structs from `size_t` to `uint32_t`

The last item is so that we have consistency across platforms in how large the
field is and can use `crc_add32` safely for host-side tests as well as
device-side ones.

<commit modified to resolve conflict>

(cherry picked from commit 10807a25f84e1c367e7c01573f9b04a0ce82b184)

+++ sw/device/lib/crypto/include/datatypes.h
@@ -396,7 +396,7 @@ typedef struct otcrypto_key_config {
   // Mode for which the key usage is intended.
   otcrypto_key_mode_t key_mode;
   // Length in bytes of the unblinded form of this key.
-  size_t key_length;
+  uint32_t key_length;
   // Whether the hardware key manager should produce this key.
   // If this is set to `true`, the keyblob must be exactly 8 words long, where
   // the first word is the version and the remaining 7 words are the salt.
@@ -414,7 +414,7 @@ typedef struct otcrypto_unblinded_key {
   // Mode for which the key usage is intended.
   otcrypto_key_mode_t key_mode;
   // Key length in bytes.
-  size_t key_length;
+  uint32_t key_length;
   // Implementation specific, storage provided by caller.
   uint32_t *key;
   // Implementation specific, checksum for this struct.
@@ -428,7 +428,7 @@ typedef struct otcrypto_blinded_key {
   // Key configuration information.
   const otcrypto_key_config_t config;
   // Length of blinded key material in bytes.
-  const size_t keyblob_length;
+  const uint32_t keyblob_length;
   // Implementation specific, storage provided by caller.
   uint32_t *keyblob;
   // Implementation specific, checksum for this struct.
